@@ -7,7 +7,7 @@ issue_tracker: https://gitlab.com/testapp-system/katex_flutter/issues
 publish_to: 'none'
 
 environment:
-  sdk: ">=2.8.1 <3.0.0"
+  sdk: ">=2.18.5 <3.0.0"
 
 dependencies:
   flutter:
